refactor: cleaned up
