+++ src/cmd/apply.ts
@@ -2,19 +2,13 @@ import { mkdirSync, rmdirSync, writeFileSync } from 'fs'
 import { isIPv6 } from 'net'
 import { Argv, CommandModule } from 'yargs'
 import { $ } from 'zx'
+import { cleanupHandler, prepareEnvironment } from '../common/cli'
+import { logLevelString, OtomiDebugger, terminal } from '../common/debug'
 import { hf, hfValues } from '../common/hf'
-import { cleanupHandler, prepareEnvironment } from '../common/setup'
-import {
-  getFilename,
-  getOtomiLoadBalancerIP,
-  getParsedArgs,
-  logLevelString,
-  OtomiDebugger,
-  setParsedArgs,
-  terminal,
-} from '../common/utils'
+import { getOtomiLoadBalancerIP } from '../common/k8s'
+import { getFilename } from '../common/utils'
 import { writeValues } from '../common/values'
-import { Arguments as HelmArgs, helmOptions } from '../common/yargs-opts'
+import { Arguments as HelmArgs, getParsedArgs, helmOptions, setParsedArgs } from '../common/yargs-opts'
 import { ProcessOutputTrimmed } from '../common/zx-enhance'
 import { Arguments as DroneArgs } from './gen-drone'
