+++ src/cmd/bootstrap.test.ts
@@ -1,0 +1,7 @@
+import { expect } from 'chai'
+
+describe('Bootstrapping', () => {
+  it('Should not re-generate passwords if already existing in secret', () => {
+    expect(true).to.be.true
+  })
+})
