+++ src/cmd/encrypt.ts
@@ -1,7 +1,9 @@
 import { Argv } from 'yargs'
 import { Arguments, encrypt as encryptFunc } from '../common/crypt'
-import { prepareEnvironment } from '../common/setup'
-import { getFilename, getParsedArgs, OtomiDebugger, setParsedArgs, terminal } from '../common/utils'
+import { prepareEnvironment } from '../common/cli'
+import { getFilename } from '../common/utils'
+import { OtomiDebugger, terminal } from '../common/debug'
+import { getParsedArgs, setParsedArgs } from '../common/yargs-opts'
 
 const cmdName = getFilename(__filename)
 const debug: OtomiDebugger = terminal(cmdName)
