+++ src/cmd/files.ts
@@ -1,7 +1,9 @@
 import { Argv } from 'yargs'
+import { prepareEnvironment } from '../common/cli'
+import { OtomiDebugger, terminal } from '../common/debug'
 import { env } from '../common/envalid'
-import { prepareEnvironment } from '../common/setup'
-import { BasicArguments, getEnvFiles, getFilename, OtomiDebugger, setParsedArgs, terminal } from '../common/utils'
+import { getEnvFiles, getFilename } from '../common/utils'
+import { BasicArguments, setParsedArgs } from '../common/yargs-opts'
 
 const cmdName = getFilename(__filename)
 const debug: OtomiDebugger = terminal(cmdName)
