+++ src/cmd/score-templates.ts
@@ -2,9 +2,10 @@ import { existsSync, unlinkSync } from 'fs'
 import { Argv } from 'yargs'
 import { $, nothrow } from 'zx'
 import { hfTemplate } from '../common/hf'
-import { cleanupHandler, prepareEnvironment } from '../common/setup'
-import { getFilename, getParsedArgs, OtomiDebugger, setParsedArgs, terminal } from '../common/utils'
-import { Arguments, helmOptions } from '../common/yargs-opts'
+import { cleanupHandler, prepareEnvironment } from '../common/cli'
+import { getFilename } from '../common/utils'
+import { Arguments, getParsedArgs, helmOptions, setParsedArgs } from '../common/yargs-opts'
+import { OtomiDebugger, terminal } from '../common/debug'
 
 const cmdName = getFilename(__filename)
 const templatePath = '/tmp/template.yaml'
