+++ src/cmd/template.ts
@@ -1,8 +1,9 @@
 import { Argv } from 'yargs'
 import { hfTemplate } from '../common/hf'
-import { prepareEnvironment } from '../common/setup'
-import { getFilename, getParsedArgs, OtomiDebugger, setParsedArgs, terminal } from '../common/utils'
-import { Arguments as HelmArgs, helmOptions } from '../common/yargs-opts'
+import { prepareEnvironment } from '../common/cli'
+import { getFilename } from '../common/utils'
+import { Arguments as HelmArgs, getParsedArgs, helmOptions, setParsedArgs } from '../common/yargs-opts'
+import { OtomiDebugger, terminal } from '../common/debug'
 
 interface Arguments extends HelmArgs {
   outDir: string
