+++ src/cmd/test.ts
@@ -3,14 +3,15 @@ import { Argv } from 'yargs'
 import { $ } from 'zx'
 import { env } from '../common/envalid'
 import { hf } from '../common/hf'
-import { cleanupHandler, prepareEnvironment } from '../common/setup'
-import { getFilename, getParsedArgs, OtomiDebugger, setParsedArgs, terminal } from '../common/utils'
-import { Arguments, helmOptions } from '../common/yargs-opts'
+import { cleanupHandler, prepareEnvironment } from '../common/cli'
+import { getFilename } from '../common/utils'
+import { Arguments, getParsedArgs, helmOptions, setParsedArgs } from '../common/yargs-opts'
 import { ProcessOutputTrimmed } from '../common/zx-enhance'
 import { diff } from './diff'
 import { lint } from './lint'
 import { validateTemplates } from './validate-templates'
 import { validateValues } from './validate-values'
+import { OtomiDebugger, terminal } from '../common/debug'
 
 const cmdName = getFilename(__filename)
 const tmpFile = '/tmp/otomi/test.yaml'
