+++ src/cmd/values.ts
@@ -1,8 +1,10 @@
 import { dump } from 'js-yaml'
 import { Argv } from 'yargs'
+import { prepareEnvironment } from '../common/cli'
+import { OtomiDebugger, terminal } from '../common/debug'
 import { hfValues } from '../common/hf'
-import { prepareEnvironment } from '../common/setup'
-import { BasicArguments, getFilename, getParsedArgs, OtomiDebugger, setParsedArgs, terminal } from '../common/utils'
+import { getFilename } from '../common/utils'
+import { BasicArguments, getParsedArgs, setParsedArgs } from '../common/yargs-opts'
 
 const cmdName = getFilename(__filename)
 const debug: OtomiDebugger = terminal(cmdName)
