+++ src/common/cli.ts
@@ -1,0 +1,54 @@
+import { readdirSync } from 'fs'
+import { chalk } from 'zx'
+import { decrypt } from './crypt'
+import { terminal } from './debug'
+import { env, isCli } from './envalid'
+import { checkKubeContext } from './k8s'
+import { isCore } from './utils'
+
+chalk.level = 2
+/**
+ * Check the ENV_DIR parameter and whether or not the folder is populated
+ * @returns
+ */
+const isReadyEnvDir = (): boolean => {
+  const d = terminal('checkEnvDir')
+  if (isCore && !env.ENV_DIR) {
+    throw new Error('The ENV_DIR environment variable is not set')
+  }
+  d.debug(`ENV_DIR: ${env.ENV_DIR}`)
+  return readdirSync(env.ENV_DIR).length > 0
+}
+
+type PrepareEnvironmentOptions = {
+  skipEnvDirCheck?: boolean
+  skipKubeContextCheck?: boolean
+  skipDecrypt?: boolean
+  skipAllPreChecks?: boolean
+}
+
+export const scriptName = process.env.OTOMI_CALLER_COMMAND ?? 'otomi'
+
+/**
+ * Prepare environment when running an otomi command
+ */
+export const prepareEnvironment = async (options?: PrepareEnvironmentOptions): Promise<void> => {
+  if (options?.skipAllPreChecks) return
+  const d = terminal('prepareEnvironment')
+  d.info('Checking environment')
+  if (!options?.skipEnvDirCheck && isReadyEnvDir()) {
+    if (isCli && !options?.skipKubeContextCheck) await checkKubeContext()
+    if (!options?.skipDecrypt) await decrypt()
+  }
+}
+
+/**
+ * Cleanup trap on exit - any handler function MUST be synchronous
+ * @param handler cleanup function set per command
+ */
+export const cleanupHandler = (handler: () => any): void => {
+  process.on('exit', (code) => {
+    handler()
+    process.exit(code)
+  })
+}
