+++ src/common/zx-enhance.ts
@@ -1,6 +1,10 @@
-import { chalk, ProcessOutput, ProcessPromise, question } from 'zx'
+import { $, chalk, ProcessOutput, ProcessPromise, question } from 'zx'
+import { DebugStream } from './debug'
 import { isCli } from './envalid'
-import { DebugStream, getParsedArgs } from './utils'
+import { getParsedArgs } from './yargs-opts'
+
+$.verbose = false // https://github.com/google/zx#verbose - don't need to print the SHELL executed commands
+$.prefix = 'set -euo pipefail;' // https://github.com/google/zx/blob/main/index.mjs#L103
 
 const MAX_RETRIES_QUESTION = 3
